chore: Remove Register component and update Event page styling
This commit removes the unused Register component and updates the styling of the Event page. The border color of the details section has been changed to #767B91, and the width of the details section has been reduced to 30vw. Additionally, the background color of the body has been set to #FCFCFC, and the width of the SVG elements has been adjusted to 10vw. The CSS for the CreateEvent page has also been modified to resize the textarea and input fields, and the width of the time input field has been set to 10vw. The price and pincode input fields have also been adjusted to have a width of 10vw. Lastly, the color of the icons has been changed to #767B91, and the focus styles for input, textarea, and select elements have been updated to remove the resize handle.