chore: Update CSS files and add popup functionality
This commit updates the CSS files for the Home, RequestDetailsPopup, and Userboard pages. It adds a transition effect to the home button in the Home page, sets the background color for the popup button in the RequestDetailsPopup page, and adds styling for the event and popup-overlay classes in the Userboard page. Additionally, it adds new routes for declining and accepting requests in the backend index.js file. These changes improve the user interface and functionality of the application.