feat: Add react-spinners package for loading animations
This commit adds the react-spinners package to the frontend dependencies in the package.json and package-lock.json files. The react-spinners package provides a collection of loading animations that can be used to enhance the user experience. By including this package, we can easily incorporate loading animations into our application when necessary.